requirements.txt
huggingface_hub
python-dotenv
PyYAML